Improve code health in src/vdp/mod.rs
- Verified that the reported duplicate `is_dma_fill` method does not exist.
- Removed unnecessary `as u16` casts identified by clippy in `src/vdp/mod.rs` to improve code health.
- Verified changes with `cargo clippy` and `cargo test`.

+++ src/vdp/mod.rs
@@ -936,7 +936,7 @@ impl Vdp {
                 let h_flip = (entry & 0x0800) != 0;
                 let tile_index = entry & 0x07FF;
 
-                let patterns = self.fetch_tile_pattern(tile_index, pixel_v as u16, v_flip);
+                let patterns = self.fetch_tile_pattern(tile_index, pixel_v, v_flip);
 
                 for i in 0..pixels_to_process {
                     let current_pixel_h = pixel_h + i;
@@ -971,7 +971,7 @@ impl Vdp {
             let h_flip = (entry & 0x0800) != 0;
             let tile_index = entry & 0x07FF;
 
-            self.draw_full_tile_row(tile_index, palette, v_flip, h_flip, pixel_v as u16, line_offset + screen_x as usize);
+            self.draw_full_tile_row(tile_index, palette, v_flip, h_flip, pixel_v, line_offset + screen_x as usize);
 
             screen_x += 8;
             scrolled_h = scrolled_h.wrapping_add(8);
@@ -993,7 +993,7 @@ impl Vdp {
                 let h_flip = (entry & 0x0800) != 0;
                 let tile_index = entry & 0x07FF;
 
-                let patterns = self.fetch_tile_pattern(tile_index, pixel_v as u16, v_flip);
+                let patterns = self.fetch_tile_pattern(tile_index, pixel_v, v_flip);
 
                 for i in 0..pixels_to_process {
                     let current_pixel_h = pixel_h + i;
